tools/docs-indexer/requirements.txt
pyyaml==6.0.1
gitpython==3.1.40
